surfman: get_monitor_info() may fail.
Don't segv when that happens, just move on.
This can happen in a variety of scenari, either the plugin saw a new
monitor but something went bad and it can't fetch its info, or the
monitor was quickly plugged and yanked, so the plugin no longer has info
on it.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>